pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastfoodgo"
version = "0.1.0"
authors = [
  { name="TonNom", email="ton@email.com" },
]
description = "FastFoodGo - Projet DevOps Data & IA"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
